Closes #514 (also code cleanup)

+++ src/main/java/shadows/apotheosis/ench/table/ApothEnchantBlock.java
@@ -93,7 +93,7 @@ public class ApothEnchantBlock extends EnchantmentTableBlock implements IReplace
 								break;
 							}
 
-							level.addParticle(ParticleTypes.ENCHANT, (double) pos.getX() + 0.5D, (double) pos.getY() + 2.0D, (double) pos.getZ() + 0.5D, (double) ((float) i + rand.nextFloat()) - 0.5D, (double) ((float) k - rand.nextFloat() - 1.0F), (double) ((float) j + rand.nextFloat()) - 0.5D);
+							level.addParticle(ParticleTypes.ENCHANT, pos.getX() + 0.5D, pos.getY() + 2.0D, pos.getZ() + 0.5D, i + rand.nextFloat() - 0.5D, k - rand.nextFloat() - 1.0F, j + rand.nextFloat() - 0.5D);
 						}
 					}
 				}
